Cleanup (+10 squashed commit)
Squashed commit:

[16ad4fe] Update the render code with the newly added classes

[32dfe36] Use logging utility instead of cout

[624f567] Fix warning

[82ca0a4] Add math utilities and definitions

[0e5b486] Add HittableList class so we can check hits for multiple objects in the scene

[e542a88] Add function to determine if surface is outside / inside at geometry time

[eee64ea] Implement the IHittable interface for sphere objects

[01133f2] Add an interface for entities in the scene which can be hit by rays

[74dbdcf] Simplify equation: Get rid of factor 2 by using 0.5*b

[7237193] Simplify equation: A vector dotted with itself is length squared

+++ RayTracer/Source/RayTracer/Private/Image.cpp
@@ -30,7 +30,7 @@ void Image::SetPixel(size_t x, size_t y, Color color)
 
 void Image::Save(const char* path)
 {
-    static const int bytes_per_row = GetWidth() * sizeof(PixelData);
+    static const int bytes_per_row = static_cast<int>(GetWidth() * sizeof(PixelData));
     static const int num_components = 4;    // rgba
     int result = stbi_write_png(path, static_cast<int>(GetWidth()), static_cast<int>(GetHeight()), num_components, static_cast<void*>(data_), bytes_per_row);
     assert(result != 0);
